+++ SOURCE/KERNEL/32RTOSKRNL/DRIVERS/ATA_PIIX3/ATA_PIIX3.c
@@ -4,7 +4,7 @@
 #include <STD/ASM.h>
 #include <STD/MEM.h>
 #include <STD/STRING.h>
-#include <DRIVERS/VIDEO/VBE.h>
+#include <DRIVERS/VESA/VBE.h>
 #include <CPU/PIC/PIC.h>
 
 /*
