+++ SOURCE/KERNEL/32RTOSKRNL/DRIVERS/DRIVER_INC.h
@@ -3,7 +3,7 @@
 #include <STD/ASM.h>
 #include <STD/MEM.h>
 #include <STD/STRING.h>
-#include <DRIVERS/VIDEO/VBE.h>
+#include <DRIVERS/VESA/VBE.h>
 #include <CPU/PIC/PIC.h>
 #include <CPU/ISR/ISR.h>
 #include <DRIVERS/PCI/PCI.h>
