+++ SOURCE/KERNEL/32RTOSKRNL/DRIVERS/VESA/VBE.c
